Use List[Double] everywhere for Cohere v4 embeddings

+++ common-lib/src/main/scala/com/gu/mediaservice/lib/aws/Bedrock.scala
@@ -76,14 +76,14 @@ class Bedrock(config: CommonConfig)
     }
   }
 
-  def createTextEmbedding(inputData: String)(implicit ec: ExecutionContext, logMarker: LogMarker): Future[List[Float]] = {
+  def createTextEmbedding(inputData: String)(implicit ec: ExecutionContext, logMarker: LogMarker): Future[List[Double]] = {
     val requestBody = createRequestBody(inputData)
     val bedrockFuture = Future { sendBedrockEmbeddingRequest(requestBody) }
     bedrockFuture.map { response =>
       val responseBody = response.body().asUtf8String()
       val json = Json.parse(responseBody)
       // Extract the embedding array (first element since it's an array of arrays)
-      val embedding = (json \ "embeddings" \ "float")(0).as[List[Float]]
+      val embedding = (json \ "embeddings" \ "float")(0).as[List[Double]]
       logger.info(
         logMarker,
         s"Successfully extracted text embedding. Vector size: ${embedding.size}"
